Strip out all usage of original Domain object
