+++ db/db.go
@@ -33,14 +33,6 @@ type Database interface {
 	PutHostnameScan(string, checker.HostnameResult) error
 	// Gets counts per day of hosts supporting MTA-STS adoption.
 	GetMTASTSStats() (models.TimeSeries, error)
-	// Upserts domain state.
-	PutDomain(models.Domain) error
-	// Retrieves state of a domain
-	GetDomain(string, models.DomainState) (models.Domain, error)
-	// Retrieves all domains in a particular state.
-	GetDomains(models.DomainState) ([]models.Domain, error)
-	SetStatus(string, models.DomainState) error
-	RemoveDomain(string, models.DomainState) (models.Domain, error)
 	ClearTables() error
 }
